The pointer is now captured on Windows (consistent with Microsoft's gesture handling examples) so that continuing the gesture outside of the control no longer pauses the gesture.
Implemented swipe (GestureSettings.CrossSlide) and pinch (GestureSettings.ManipulationScale), while pan is now done via GestureSettings.ManipulationTranslateX and GestureSettings.ManipulationTranslateY (consistent with Microsoft's examples - Drag is primarily intended for moving selected objects).  Pinch is particularly challenging though as it requires more work than perhaps is expected due to potential collisions between pinching and panning.  As such the recommendation is to store and process the pointer data manually to discern between pinching and panning, and then manage the state.

Swiping is all handled automatically via CrossSliding, though some steps are still taken to identify the swipe's direction.  And unfortunately on Windows, support for swipes can only be horizontal or vertical.  Not sure why.  I default it to horizontal.

UserInteraction.Alert isn't implemented on Windows and so fails with a NotImplementedException.  The exception is now caught and the alert message is outputted to the debug console as a fallback.

Adjusted the default height of the demo app so that the bottom controls aren't cut off.