Added the Windows-only CrossSlideHorizontallyProperty.  This allows an element to define whether swipes are either horizontal or vertical on Windows, since Windows only allows one or the other.
For the demo app, the appxmanifest needed to be updated (as per https://learn.microsoft.com/en-us/dotnet/communitytoolkit/maui/alerts/snackbar?tabs=windows%2Candroid) to get toasts/snackbars working fully on Windows 11.  In addition to that, the Windows app package must be an MSIX in order to ensure they work properly on Win 11.

Renamed ReturnAllPointsOnWindows to ProcessIntermediatePoints.  Up for debate if you don't want to change a previously published API.  I just find this is more descriptive.